+++ installer/db/mysql/mediaexportdb/init_01_create-database.sql
@@ -1,6 +1,0 @@
--- ------------------------------------
--- create current Database: `mediaexportdb`
--- ------------------------------------
-CREATE DATABASE IF NOT EXISTS mediaexportdb DEFAULT CHARACTER SET latin1 COLLATE latin1_general_ci;
-
-USE `mediaexportdb`;