+++ installer/db/mysql/mytbdb/init_01_create-database.sql
@@ -1,0 +1,5 @@
+-- ------------------------------------
+-- create current Database: `mytbdb`
+-- ------------------------------------
+CREATE DATABASE IF NOT EXISTS testmytbdb DEFAULT CHARACTER SET latin1 COLLATE latin1_general_ci;
+USE `testmytbdb`;