+++ installer/db/mysql/mytbexportdb/init_01_create-database.sql
@@ -1,0 +1,6 @@
+-- ------------------------------------
+-- create current Database: `mytbexportdb`
+-- ------------------------------------
+CREATE DATABASE IF NOT EXISTS mytbexportdb DEFAULT CHARACTER SET latin1 COLLATE latin1_general_ci;
+
+USE `mytbexportdb`;