+++ src/shared/tdoc-commons/services/tdoc-sql-mytbdb.config.ts
@@ -1,7 +1,7 @@
 import {TableConfig, TableConfigs} from '@dps/mycms-commons/dist/search-commons/services/sql-query.builder';
 import {AdapterFilterActions} from '@dps/mycms-commons/dist/search-commons/services/mapper.utils';
 
-export class TourDocSqlMediadbConfig {
+export class TourDocSqlMytbDbConfig {
     public static personCategories = ['Person', 'person', 'Familie', 'family', 'friend', 'Freund'];
     public static tableConfigs: TableConfigs = {
         'track': {
@@ -586,7 +586,7 @@ export class TourDocSqlMediadbConfig {
                     'FROM' +
                     ' objects LEFT JOIN image_object ON objects.o_key=image_object.io_obj_type' +
                     ' INNER JOIN image ON image_object.i_id=image.i_id ' +
-                    ' WHERE objects.o_category in ("' + TourDocSqlMediadbConfig.personCategories.join('", "') + '")' +
+                    ' WHERE objects.o_category in ("' + TourDocSqlMytbDbConfig.personCategories.join('", "') + '")' +
                     ' GROUP BY value' +
                     ' ORDER BY value',
                     filterField: 'o_name',
@@ -987,7 +987,7 @@ export class TourDocSqlMediadbConfig {
                         ' o_name AS value ' +
                         'FROM' +
                         ' objects LEFT JOIN image_object ON objects.o_key=image_object.io_obj_type' +
-                        ' WHERE objects.o_category in ("' + TourDocSqlMediadbConfig.personCategories.join('", "') + '")' +
+                        ' WHERE objects.o_category in ("' + TourDocSqlMytbDbConfig.personCategories.join('", "') + '")' +
                         ' GROUP BY value' +
                         ' ORDER BY value',
                     filterField: 'o_name',
@@ -1308,7 +1308,7 @@ export class TourDocSqlMediadbConfig {
                     ' o_name AS value ' +
                     'FROM' +
                     ' objects LEFT JOIN video_object ON objects.o_key=video_object.vo_obj_type' +
-                    ' WHERE objects.o_category in ("' + TourDocSqlMediadbConfig.personCategories.join('", "') + '")' +
+                    ' WHERE objects.o_category in ("' + TourDocSqlMytbDbConfig.personCategories.join('", "') + '")' +
                     ' GROUP BY value' +
                     ' ORDER BY value',
                     filterField: 'o_name',
@@ -2375,7 +2375,7 @@ export class TourDocSqlMediadbConfig {
     };
 
     public getTableConfigForTableKey(table: string): TableConfig {
-        return TourDocSqlMediadbConfig.tableConfigs[table];
+        return TourDocSqlMytbDbConfig.tableConfigs[table];
     }
 }